test: correct compile warnings
Fix several similar compile warnings:

flux-compose/test/test-vector.c:42:19: warning: format specifies type 'int' but the argument has type 'size_t' (aka 'unsigned long') [-Wformat]
  ASSERT_INT(512, test_vector->buffer_size);
  ~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~

+++ test/test.h
@@ -31,7 +31,7 @@ extern char fail_message[2048];
 
 #define ASSERT_INT(expected, actual)                                            \
   if (actual != expected) {                                                     \
-    FAIL("Expected integer: %d\n                   got: %d\n               at line: %d\n", expected, actual, __LINE__); \
+    FAIL("Expected integer: %ld\n                   got: %ld\n               at line: %d\n", (long int) expected, (long int) actual, __LINE__); \
   }
 
 void test_vector_suite();
